Fix logging line and call method information
The logging line number and file was for the previous call position, not for
where the actual log entry was called

Also fix for ErrorMessage class calls with shifting the start position up depending on which method is called.

Output shows file and line where the message/log call was done and the function/class method where the log call was done

+++ www/lib/CoreLibs/Convert/Byte.php
@@ -77,7 +77,7 @@ class Byte
 			// labels in order of size [Y, Z]
 			$labels = ['', 'K', 'M', 'G', 'T', 'P', 'E'];
 			// exp position calculation
-			$exp = floor(log($abs_bytes, $unit));
+			$exp = (int)floor(log($abs_bytes, $unit));
 			// avoid printing out anything larger than max labels
 			if ($exp >= count($labels)) {
 				$exp = count($labels) - 1;
